move to src, add JSON export support
