+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "github-projects-scripts",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "src/index.js",
   "type": "module",
   "license": "MIT",
   "dependencies": {
